requirements.txt
opencv-python==4.10.0
pytorch3d
open3d
tensorflow==2.15.0
plyfile
opencv-python
lpips
trimesh
torch-kmeans
tqdm
diffusers>=0.32.1
accelerate>=1.1.1
transformers>=4.46.2
imageio>=2.35.1
imageio-ffmpeg>=0.5.1
openai>=1.54.0
moviepy>=2.0.0
scikit-video>=1.1.11
pydantic>=2.10.3
git+https://github.com/diqihe666/diff-langsurf-rasterizer.git
git+https://github.com/openai/CLIP.git